# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "contact": {
            "email": "contact@touri-touri.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/site-dates": {
            "get": {
                "tags": [
                    "Site Dates"
                ],
                "summary": "Get list of site dates",
                "description": "Returns list of site dates",
                "operationId": "getSiteDatesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SiteDate"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Site Dates"
                ],
                "summary": "Store new site date",
                "description": "Returns site date data",
                "operationId": "storeSiteDate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSiteDateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SiteDate"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/site-dates/{id}": {
            "get": {
                "tags": [
                    "Site Dates"
                ],
                "summary": "Get site date information",
                "description": "Returns site date data",
                "operationId": "getSiteDateById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of site date to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SiteDate"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Site Dates"
                ],
                "summary": "Update existing site date",
                "description": "Returns updated site date data",
                "operationId": "updateSiteDate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of site date to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSiteDateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SiteDate"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Site Dates"
                ],
                "summary": "Delete existing site date",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteSiteDate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of site date to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SiteDate"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SiteDate": {
                "title": "SiteDate",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "site_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "date": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "2021-05-05"
                    },
                    "created_at": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "2021-05-05T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "2021-05-05T12:00:00.000000Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "2021-05-05T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "StoreSiteDateRequest": {
                "title": "StoreSiteDateRequest",
                "properties": {
                    "site_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "date": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "2021-05-05"
                    }
                },
                "type": "object"
            },
            "UpdateSiteDateRequest": {
                "title": "UpdateSiteDateRequest",
                "properties": {
                    "site_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "date": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "2021-05-05"
                    }
                },
                "type": "object"
            }
        }
    }
}